requirements.txt
aioredis==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
celery==5.2.7
channels==3.0.5
daphne==3.0.2
Django==4.1
django-cors-headers==3.13.0
django-import-export==2.8.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hiredis==2.0.0
Markdown==3.4.1
phonenumbers==8.12.52
Pillow==9.2.0
requests==2.28.1
requests-oauthlib==1.3.1